fix : 코드카타 알고리즘 16. x만큼 간격이 있는 n개의 숫자 - 리스트 범위는 long

+++ src/algorithm/SequenceWithStepX.java
@@ -4,7 +4,7 @@ public class SequenceWithStepX {
     public long[] SequenceWithStepX(int x, int n) {
         long[] answer = new long[n];
         for (int i = 0; i < n; i++) {
-            answer[i] = x * (i + 1);
+            answer[i] = (long) x * (i + 1);
         }
 
         return answer;
